some @mentions in preview
replaces immediatly if there is one match,
but just spews multiple ones.

idea is to have <option>s if there are multiple
and sending that back kind of like a dictionary for the next replace
making the lookup from name to feed explicit

* add silly in-memory about:name cache

rational is also commented the code. the tl:dr; is oasis already doesnt
use ssb-about or other indexing plugins.

butt for @mentions we need a quick way to lookup names to feeds,
otherwise we block the preview flow with long query times.

* haunted: sometimes i see an avatar image, sometimes i dont..!?